backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.0
sqlalchemy==2.0.23
lxml==4.9.3

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# Email
python-smtplib

# Scheduling
apscheduler==3.10.4

# Your existing scraping dependencies
selenium==4.15.2
beautifulsoup4==4.12.2
webdriver-manager==4.0.1
requests==2.31.0